requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.1.4
asgiref==3.8.1
attrs==23.2.0
Babel==2.10.3
bcc==0.1.10
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
CacheControl==0.14.0
certifi==2023.11.17
cffi==1.17.0
chardet==5.2.0
click==8.1.6
cloud-init==24.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
configparser==7.1.0
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
customtkinter==5.2.2
darkdetect==0.8.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==5.1
djangorestframework==3.15.2
duplicity==2.1.4
fasteners==0.18
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.2
idna==3.6
imagetk==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.5.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
MouseInfo==0.1.3
msgpack==1.0.3
netaddr==0.8.0
netifaces==0.11.0
numpy==2.1.0
oauthlib==3.2.2
olefile==0.46
openai==1.43.0
packaging==24.0
pandas==2.2.2
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
psutil==5.9.8
ptyprocess==0.7.0
PyAutoGUI==0.9.54
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.1
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyRect==0.2.0
pyrsistent==0.20.0
PyScreeze==1.0.1
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.2.2
requests==2.31.0
rich==13.7.1
scipy==1.14.1
scour==0.38.2
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
ssh-import-id==5.11
systemd-python==235
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.42.0
xdg==5
xkit==0.0.0
